backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
pydantic==2.5.1
python-dotenv==1.0.0
httpx==0.25.1
requests==2.31.0
google-api-python-client==2.97.0
yt-dlp==2023.3.4
ffmpeg-python==0.2.0
